refactor(correlation): move upstream correlation under agent
