+++ app/agent/correlation/providers.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from collections.abc import Callable
 
-from app.correlation.upstream import (
+from app.agent.correlation.upstream import (
     LogSignal,
     MetricSeries,
     TopologyHint,
