+++ tests/synthetic/rds_postgres/correlation/candidate_scoring.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.scoring import (
+from app.agent.correlation.scoring import (
     CandidateCorrelationScore,
     score_candidate_correlation,
 )
