+++ tests/synthetic/rds_postgres/correlation/investigation_flow.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.upstream import MetricSeries, UpstreamEvidenceBundle
+from app.agent.correlation.upstream import MetricSeries, UpstreamEvidenceBundle
 from tests.synthetic.rds_postgres.correlation.candidate_scoring import (
     score_candidate_correlation,
 )
