+++ tests/synthetic/rds_postgres/correlation/models.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.models import CorrelatedSignal, UpstreamCandidate
+from app.agent.correlation.models import CorrelatedSignal, UpstreamCandidate
 
 __all__ = [
     "CorrelatedSignal",
