+++ tests/synthetic/rds_postgres/correlation/periodicity.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.scoring import (
+from app.agent.correlation.scoring import (
     PeriodicityScore,
     score_periodic_spikes,
 )
