+++ tests/synthetic/rds_postgres/correlation/ranking.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.scoring import rank_upstream_candidates
+from app.agent.correlation.scoring import rank_upstream_candidates
 
 __all__ = [
     "rank_upstream_candidates",
