+++ tests/synthetic/rds_postgres/correlation/reporting.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.reporting import (
+from app.agent.correlation.reporting import (
     CorrelationReport,
     build_correlation_report,
     correlation_report_to_payload,
