+++ tests/synthetic/rds_postgres/correlation/topology.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from app.correlation.scoring import (
+from app.agent.correlation.scoring import (
     TopologyCorrelation,
     TopologyNode,
     score_topology_adjacency,
