+++ tests/test_datadog_provider.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from app.correlation.datadog_adapter import DatadogCorrelationAdapter
-from app.correlation.datadog_provider import (
+from app.agent.correlation.datadog_adapter import DatadogCorrelationAdapter
+from app.agent.correlation.datadog_provider import (
     DatadogCorrelationQueries,
     DatadogUpstreamEvidenceProvider,
 )
