+++ tests/test_upstream_correlation_provider.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
-from app.correlation.providers import (
+from app.agent.correlation.providers import (
     NoopUpstreamEvidenceProvider,
     QueryBackedUpstreamEvidenceProvider,
 )
-from app.correlation.upstream import (
+from app.agent.correlation.upstream import (
     LogSignal,
     MetricSeries,
     TopologyHint,
